refactor: merge resource diagram mix tasks
